demo: Extract the Organization JSON-LD node builder
The Article structured data (added next) needs the site's `Organization` as a
nested `author`/`publisher` node, i.e. the same mapping from the site config that
`OrganizationJsonLd` already does, but without the `@context` wrapper.

Move the mapping into `buildOrganizationNode`, add a context-wrapped
`buildOrganization` for use at the page root, and have `OrganizationJsonLd`
render that. No change to the emitted output.
